Try3 - supporting dfsan_get_label
In order to support this API we must compile tsan with dfsan. We can't use -santize=dataflow because used LLVM is 14 and we need an older dfsan to support all labels.
So, we spliting the tsan and dfsan passes, and compile tsan with dfsan.

We need to fix some things for this:
* Disable the data race detection in tsan - we can't use the shadow memory in tsan because the same region is used in dfsan. So we disable it.
* Supporting some symbols problems - like ASM symbols
* updating build script